chore: rename to deprecated
